+++ drive/src/drive/fee_pools/epoch/mod.rs
@@ -1,4 +1,3 @@
-pub mod tree_key_constants;
 pub mod epoch_pool;
-pub mod fees;
+pub mod credit_distribution_pools;
 pub mod proposers;
