+++ drive/src/drive/fee_pools/mod.rs
@@ -1,7 +1,11 @@
+use crate::drive::RootTree;
+
 pub mod constants;
 pub mod epoch;
 pub mod fee_distribution;
 pub mod fee_pools;
 pub mod storage_fee_distribution_pool;
-#[cfg(test)]
-pub mod tests;
+
+pub(crate) fn fee_pool_vec_path() -> Vec<Vec<u8>> {
+    vec![vec![RootTree::Pools as u8]]
+}
