--- drive/src/drive/fee_pools/tests/helpers/mn_shares_contract.rs
@@ -1,1 +1,0 @@
-
