+++ drive/src/fee_pools/epoch_pool/mod.rs
@@ -1,11 +1,6 @@
-use std::borrow::Cow;
-use grovedb::batch::{GroveDbOp, Op};
-use grovedb::{Element, TransactionArg};
-use grovedb::batch::Op::Insert;
-use crate::drive::batch::GroveDbOpBatch;
-use crate::drive::fee_pools::epoch::tree_key_constants;
-use crate::drive::fee_pools::epoch::tree_key_constants::{KEY_FEE_MULTIPLIER, KEY_POOL_PROCESSING_FEES, KEY_POOL_STORAGE_FEES, KEY_START_BLOCK_HEIGHT, KEY_START_TIME};
-use crate::error::Error;
+pub mod paths;
+pub mod operations_factory;
+pub mod tree_key_constants;
 
 pub struct EpochPool {
     pub index: u16,
@@ -18,221 +13,6 @@ impl EpochPool {
             index,
             key: index.to_be_bytes(),
         }
-    }
-}
-
-impl EpochPool {
-    pub fn add_shift_current_epoch_pool_operations(
-        &self,
-        current_epoch_pool: &EpochPool,
-        start_block_height: u64,
-        start_block_time_ms: u64,
-        fee_multiplier: u64,
-        batch: &mut GroveDbOpBatch,
-    ) {
-        // create and init next thousandth epoch
-        let next_thousandth_epoch = EpochPool::new(current_epoch_pool.index + 1000);
-        next_thousandth_epoch.add_init_empty_operations(batch);
-
-        // init first_proposer_block_height and processing_fee for an epoch
-        current_epoch_pool.add_init_current_operations(
-            fee_multiplier,
-            start_block_height,
-            start_block_time_ms,
-            batch,
-        );
-    }
-
-
-    pub fn add_init_empty_operations(&self, batch: &mut GroveDbOpBatch) {
-        batch.insert_empty_tree(FeePools::get_path(), self.key.to_vec());
-
-        // init storage fee item to 0
-        batch.push(self.update_storage_fee_operation( 0));
-    }
-
-    pub fn add_init_current_operations(
-        &self,
-        multiplier: u64,
-        start_block_height: u64,
-        start_time_ms: u64,
-        batch: &mut GroveDbOpBatch,
-    ) {
-        batch.push(self.update_start_block_height_operation(start_block_height));
-
-        batch.push(self.update_processing_fee_operation(0u64));
-
-        batch.push(self.init_proposers_operation());
-
-        batch.push(self.update_fee_multiplier_operation(multiplier));
-
-        batch.push(self.update_start_time_operation(start_time_ms));
-    }
-
-    pub fn add_mark_as_paid_operations(
-        &self,
-        transaction: TransactionArg,
-        batch: &mut GroveDbOpBatch,
-    ) -> Result<(), Error> {
-        self.add_delete_proposers_tree_operations(batch, transaction)?;
-
-        self.add_delete_storage_fee_operations(batch, transaction)?;
-
-        self.add_delete_processing_fee_operations(batch, transaction)?;
-
-        Ok(())
-    }
-
-    pub fn get_path(&self) -> [&[u8]; 2] {
-        [FeePools::get_path()[0], &self.key]
-    }
-
-    pub fn get_vec_path(&self) -> Vec<Vec<u8>> {
-        vec![ FEE, self.key.to_vec()]
-    }
-
-    pub fn update_start_time_operation(
-        &self,
-        time_ms: u64,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: KEY_START_TIME.to_vec(),
-            op: Insert {
-                element : Element::Item(time_ms.to_be_bytes().to_vec(), None)
-            }
-        }
-    }
-
-    pub fn update_start_block_height_operation(
-        &self,
-        start_block_height: u64,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: KEY_START_BLOCK_HEIGHT.to_vec(),
-            op: Insert {
-                element : Element::Item(start_block_height.to_be_bytes().to_vec(), None)
-            }
-        }
-    }
-
-    pub fn update_fee_multiplier_operation(
-        &self,
-        multiplier: u64,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path,
-            key: KEY_FEE_MULTIPLIER.to_vec(),
-            op: Insert { element: Element::Item(multiplier.to_be_bytes().to_vec(), None)}
-        }
-    }
-
-    pub fn update_processing_fee_operation(
-        &self,
-        processing_fee: u64,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: KEY_POOL_PROCESSING_FEES.to_vec(),
-            op: Insert { element: Element::new_item(processing_fee.to_be_bytes().to_vec())}
-        }
-    }
-
-    pub fn delete_processing_fee_operation(
-        &self,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: KEY_POOL_PROCESSING_FEES.to_vec(),
-            op: Op::Delete
-        }
-    }
-
-    pub fn update_storage_fee_operation(
-        &self,
-        storage_fee: u64,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: KEY_POOL_STORAGE_FEES.to_vec(),
-            op: Insert { element: Element::new_item(storage_fee.to_be_bytes().to_vec())}
-        }
-    }
-
-    pub fn delete_storage_fee_operation(
-        &self
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: KEY_POOL_STORAGE_FEES.to_vec(),
-            op: Op::Delete
-        }
-    }
-
-    fn update_proposer_block_count_operation(
-        &self,
-        proposer_pro_tx_hash: &[u8; 32],
-        block_count: u64,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_proposers_path(),
-            key: proposer_pro_tx_hash.to_vec(),
-            op: Insert { element : Element::Item(block_count.to_be_bytes().to_vec(), None)}
-        }
-    }
-
-    pub fn add_increment_proposer_block_count_operations(
-        &self,
-        proposer_pro_tx_hash: &[u8; 32],
-        transaction: TransactionArg,
-        batch: &mut GroveDbOpBatch,
-    ) -> Result<(), Error> {
-        // update proposer's block count
-        let proposed_block_count = self
-            .get_proposer_block_count(proposer_pro_tx_hash, transaction)
-            .or_else(|e| match e {
-                Error::GroveDB(grovedb::Error::PathKeyNotFound(_)) => Ok(0u64),
-                _ => Err(e),
-            })?;
-
-        let op = self.update_proposer_block_count_operation(
-            proposer_pro_tx_hash,
-            proposed_block_count + 1,
-        );
-
-        batch.push(op);
-
-        Ok(())
-    }
-
-    pub fn init_proposers_operation(&self) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: tree_key_constants::KEY_PROPOSERS.to_vec(),
-            op: Op::Insert { element: Element::empty_tree()}
-        }
-    }
-
-    pub fn delete_proposers_tree_operation(
-        &self,
-    ) -> GroveDbOp {
-        GroveDbOp {
-            path: self.get_vec_path(),
-            key: tree_key_constants::KEY_PROPOSERS.to_vec(),
-            op: Op::Delete
-        }
-    }
-
-    pub fn add_delete_proposers_operations(
-        &self,
-        pro_tx_hashes: Vec<Vec<u8>>,
-        batch: &mut GroveDbOpBatch,
-    ) -> Result<(), Error> {
-        for pro_tx_hash in pro_tx_hashes.into_iter() {
-            batch.delete(self.get_proposers_path(), Cow::from(pro_tx_hash))?;
-        }
-        Ok(())
     }
 }
